# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Care Commons API",
    "version": "1.0.0",
...
    "contact": {
      "name": "Care Commons Team",
      "url": "https://github.com/neighborhood-lab/care-commons",
      "email": "support@carecommons.example"
    },
    "license": {
      "name": "AGPL-3.0",
      "url": "https://www.gnu.org/licenses/agpl-3.0.en.html"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    },
    {
      "url": "https://preview.carecommons.example",
      "description": "Staging server"
    },
    {
      "url": "https://care-commons.vercel.app",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from login endpoint"
      }
    },
    "schemas": {
...
      "CarePlan": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Care plan unique identifier"
          },
          "clientId": {
            "type": "string",
            "format": "uuid",
            "description": "Associated client ID"
          },
          "title": {
            "type": "string",
            "description": "Care plan title"
          },
          "description": {
            "type": "string",
            "description": "Care plan description"
          },
          "startDate": {
            "type": "string",
            "format": "date",
            "description": "Care plan start date"
          },
          "endDate": {
            "type": "string",
            "format": "date",
            "description": "Care plan end date"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "completed",
              "cancelled"
            ],
            "description": "Care plan status"
          },
          "tasks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Task"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Task": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Task unique identifier"
          },
          "carePlanId": {
            "type": "string",
            "format": "uuid",
            "description": "Associated care plan ID"
          },
          "title": {
            "type": "string",
            "description": "Task title"
          },
          "description": {
            "type": "string",
            "description": "Task description"
          },
          "frequency": {
            "type": "string",
            "enum": [
              "daily",
              "weekly",
              "monthly",
              "as_needed"
            ],
            "description": "Task frequency"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "in_progress",
              "completed"
            ],
            "description": "Task status"
          },
          "dueDate": {
            "type": "string",
            "format": "date-time",
            "description": "Task due date"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "PaginatedResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "pagination": {
            "type": "object",
            "properties": {
              "page": {
                "type": "integer"
              },
              "limit": {
                "type": "integer"
              },
              "total": {
                "type": "integer"
              },
              "totalPages": {
                "type": "integer"
              }
            }
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/care-plans": {
      "post": {
        "tags": [
          "Care Plans"
        ],
        "summary": "Create new care plan",
        "description": "Create a new care plan for a client",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "clientId",
                  "title",
                  "planType"
                ],
                "properties": {
                  "clientId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "title": {
                    "type": "string",
                    "example": "Personal Care Plan"
                  },
                  "description": {
                    "type": "string"
                  },
                  "planType": {
                    "type": "string",
                    "enum": [
                      "PERSONAL_CARE",
                      "COMPANION",
                      "SKILLED_NURSING",
                      "THERAPY",
                      "HOSPICE",
                      "RESPITE",
                      "LIVE_IN",
                      "CUSTOM"
                    ]
                  },
                  "startDate": {
                    "type": "string",
                    "format": "date"
                  },
                  "endDate": {
                    "type": "string",
                    "format": "date"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Care plan created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CarePlan"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Not authenticated"
          },
          "500": {
            "description": "Server error"
          }
        }
      },
      "get": {
        "tags": [
          "Care Plans"
        ],
        "summary": "Search care plans",
        "description": "Search and filter care plans with pagination",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "query",
            "schema": {
              "type": "string"
            },
            "description": "Search query text"
          },
          {
            "in": "query",
            "name": "clientId",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Filter by client ID"
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": [
                "DRAFT",
                "PENDING_APPROVAL",
                "ACTIVE",
                "ON_HOLD",
                "EXPIRED",
                "DISCONTINUED",
                "COMPLETED"
              ]
            },
            "description": "Filter by status (comma-separated for multiple)"
          },
          {
            "in": "query",
            "name": "planType",
            "schema": {
              "type": "string",
              "enum": [
                "PERSONAL_CARE",
                "COMPANION",
                "SKILLED_NURSING",
                "THERAPY",
                "HOSPICE",
                "RESPITE",
                "LIVE_IN",
                "CUSTOM"
              ]
            },
            "description": "Filter by plan type (comma-separated for multiple)"
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            },
            "description": "Page number"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 20
            },
            "description": "Items per page"
          }
        ],
        "responses": {
          "200": {
            "description": "Care plans retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/api/care-plans/{id}": {
      "get": {
        "tags": [
          "Care Plans"
        ],
        "summary": "Get care plan by ID",
        "description": "Retrieve a single care plan by its unique identifier",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Care plan UUID"
          }
        ],
        "responses": {
          "200": {
            "description": "Care plan found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CarePlan"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "404": {
            "description": "Care plan not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
...
  }
}